feat(backend): Add endpoint to list quizzes

+++ src/main/java/com/upc/talkia_proyect/entities/User.java
@@ -36,7 +36,7 @@ public class User {
     private LocalDate dateOfBirth;
 
     @Column(name = "total_points")
-    private Integer totalPoints;
+    private Double totalPoints;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "roles_id")
